+++ R/mayoshinytheme.R
@@ -7,6 +7,7 @@
 #'     page-builder functions, e.g., \code{\link[shiny]{fluidPage}}.
 #' @return Returns a file path reference to the Mayo Shiny CSS file,
 #'     and attaches that path so shiny knows about it.
+#' @param theme Name of a theme.
 #' @seealso \code{\link[shinythemes]{shinytheme}}
 #' @examples
 #' \dontrun{
@@ -16,7 +17,16 @@
 #' )
 #' }
 #' @export
-mayoshinytheme <- function() {
+mayoshinytheme <- function(theme = c("classic", "horizon")) {
+
+  theme <- match.arg(theme)
+
+  if (theme == "horizon") {
+    file <- "bootstrap_horizon.css"
+  } else {
+    file <- "bootstrap.css"
+  }
+
   shiny::addResourcePath("mayoshinythemes", mayoshinypath())
-  file.path("mayoshinythemes", "css", "bootstrap.css")
+  file.path("mayoshinythemes", "css", file)
 }